requirements.txt
arabic-reshaper==3.0.0
asgiref==3.8.1
asn1crypto==1.5.1
Brotli==1.1.0
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.0
click==8.1.7
cryptography==43.0.3
cssselect2==0.7.0
Django==5.1.2
django-environ==0.11.2
django-filter==24.3
djangorestframework==3.15.2
et-xmlfile==1.1.0
fonttools==4.54.1
html5lib==1.1
idna==3.10
lxml==5.3.0
Markdown==3.7
openpyxl==3.1.5
oscrypto==1.3.0
pillow==11.0.0
psycopg2-binary==2.9.10
pycparser==2.22
pydyf==0.11.0
pyHanko==0.25.1
pyhanko-certvalidator==0.26.3
pypdf==5.0.1
pyphen==0.16.0
python-bidi==0.6.3
PyYAML==6.0.2
qrcode==8.0
reportlab==4.2.5
requests==2.32.3
six==1.16.0
sqlparse==0.5.1
svglib==1.5.1
tinycss2==1.3.0
tzlocal==5.2
uritools==4.0.3
urllib3==2.2.3
weasyprint==62.3
webencodings==0.5.1
xhtml2pdf==0.2.16
zopfli==0.2.3.post1
